% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create Map Label}
\usage{
eq_create_label(filtered_data)
}
\arguments{
\item{filtered_data}{DataFrame with information about earthquakes, must have LOCATION,
EQ_PRIMARY, and TOTAL_DEATHS columns}
}
\value{
Vector with content to be used by eq_map function.
}
\description{
This function creates content to be displayed inside of the popups plotted by the eq_map function.
}
\examples{

\dontrun{eq_create_label(.)}
}
