% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Data (Location)}
\usage{
eq_location_clean(clean_data)
}
\arguments{
\item{clean_data}{A dataframe object returned by the 
\code{eq_clean_data} function}
}
\value{
Adds a new column to the DataFrame called LOCATION, following the specifications:
        Title Case and without country name.
}
\description{
This function is a intermediate step to clean the raw data. The `eq_clean_data` will use it behind
the scenes to insert the column LOCATION.
}
\examples{
\dontrun{eq_location_clean(clean_data)}
}
