% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Create Map}
\usage{
eq_map(filtered_data, annot_col = "DATE")
}
\arguments{
\item{filtered_data}{A DataFrame with DATE, LONGITUDE, LATITUDE, and EQ_PRIMARY columns.}

\item{annot_col}{The information to be displayed inside of the popup.}
}
\value{
A leaflet object representing the map of earthquake points
}
\description{
This function plots a straightforward map using circles to determine the eartquake location. The radius
of the circle is according to the magnitude (in Richter Scale) of the earthquake.
}
\examples{
\dontrun{readr::read_delim("./inst/extdata/dataset.txt", delim = "\t") \%>\% 
  eq_clean_data() \%>\% 
  eq_location_clean() \%>\%
  filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% 
  mutate(popup_text = eq_create_label(.)) \%>\% 
  eq_map(annot_col = "popup_text")}
}
